requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
zipp==3.8.1
